repo-intel-bundle-v2/pyproject.toml
[project]
name = "repo-intel-bundle-v2"
version = "0.1.0"
description = "MCP server for static analysis of GitHub repositories (v2)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=0.5.0",
    "requests",
    "python-dotenv",
    "pydantic",
    "setuptools",
    "wheel",
    "jinja2",
    "fastapi",
    "uvicorn"
]

[project.scripts]
repo-intel = "repo_intel_bundle_v2.server:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
ui = [
    "streamlit",
    "graphviz",
    "pandas",
    "watchdog",
    "plotly"
]
